CARDS-2654: Refactor Group and User related components to functional Renamed login components
